backend/requirements.txt
fastapi==0.117.1
PyMuPDF==1.24.14
fuzzywuzzy==0.18.0
Jinja2==3.1.6
nltk==3.9.1
numpy==2.3.3
pydantic==2.11.9
python-dotenv==1.1.1
python-multipart==0.0.20
rapidfuzz==3.14.1
Requests==2.32.5
scikit_learn==1.7.2
uvicorn==0.3
httpx==0.27.0